Refactor: Convert App.js into multiple reusable components
