+++ src/utils/translations.js
@@ -1,0 +1,56 @@
+export const translations = {
+    english: {
+        title: "VS Code Notes",
+        placeholder: "Write your thoughts...",
+        saveButton: "Save Note",
+        clearButton: "Clear All",
+        saveNote: "Saved Notes",
+        languageLabel: "Language:",
+        pin: "Pin",
+        jsonValidator: "JSON Validator & Formatter",
+        jsonPlaceholder: "Paste or type your JSON here...",
+        validateButton: "Validate JSON",
+        beautifyButton: "Beautify JSON",
+        clearJsonButton: "Clear JSON",
+        validJson: "✅ Valid JSON",
+        invalidJson: "❌ Invalid JSON:",
+        beautifiedJson: "Beautified JSON:",
+        copyJsonButton: "Copy JSON",
+    },
+    hindi: {
+        title: "VS कोड नोट्स",
+        placeholder: "अपने विचार लिखें...",
+        saveButton: "नोट सहेजें",
+        clearButton: "सब साफ करें",
+        saveNote: "सहेजे गए नोट्स",
+        languageLabel: "भाषा:",
+        pin: "पिन करें",
+        jsonValidator: "JSON सत्यापनकर्ता और फॉर्मेटर",
+        jsonPlaceholder: "अपना JSON यहाँ पेस्ट या टाइप करें...",
+        validateButton: "JSON सत्यापित करें",
+        beautifyButton: "JSON को सुंदर बनाएं",
+        clearJsonButton: "JSON साफ करें",
+        validJson: "✅ वैध JSON",
+        invalidJson: "❌ अवैध JSON:",
+        beautifiedJson: "सुंदर JSON:",
+        copyJsonButton: "JSON कॉपी करें",
+    },
+    spanish: {
+        title: "Notas de VS Code",
+        placeholder: "Escribe tus pensamientos...",
+        saveButton: "Guardar Nota",
+        clearButton: "Borrar Todo",
+        saveNote: "Notas guardadas",
+        languageLabel: "Idioma:",
+        pin: "Fijar",
+        jsonValidator: "Validador y Formateador JSON",
+        jsonPlaceholder: "Pega o escribe tu JSON aquí...",
+        validateButton: "Validar JSON",
+        beautifyButton: "Embellecer JSON",
+        clearJsonButton: "Limpiar JSON",
+        validJson: "✅ JSON Válido",
+        invalidJson: "❌ JSON Inválido:",
+        beautifiedJson: "JSON Embellecido:",
+        copyJsonButton: "Copiar JSON",
+    },
+};
