+++ src/utils/useLocalStorage.js
@@ -1,0 +1,24 @@
+import { useState, useEffect } from 'react';
+
+export const useLocalStorage = (key, initialValue) => {
+    const [storedValue, setStoredValue] = useState(() => {
+        try {
+            const item = localStorage.getItem(key);
+            return item ? JSON.parse(item) : initialValue;
+        } catch (error) {
+            console.log("localStorage not available in VS Code context");
+            return initialValue;
+        }
+    });
+
+    const setValue = (value) => {
+        try {
+            setStoredValue(value);
+            localStorage.setItem(key, JSON.stringify(value));
+        } catch (error) {
+            console.log("localStorage not available in VS Code context");
+        }
+    };
+
+    return [storedValue, setValue];
+};
